- (snmp_api.h): version -> u_long.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@3056 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/snmp_api.h
@@ -143,7 +143,7 @@ struct snmp_session {
 	/*
 	 * Protocol-version independent fields
 	 */
-    int	    version;
+    u_long  version;
     int	    retries;	/* Number of retries before timeout. */
     long    timeout;    /* Number of uS until first timeout, then exponential backoff */
     u_long  flags;
